Move table of contents to root, rm D1S2
This finishes off the back side of disk 1 except for the TOC which hasn't been
cleaned up yet in the least.  The TOC itself is not useful to us as is because
we don't yet have any hyperlinks.  That'll be the next step--but we'll clean it
up in preparation of that anyway.